minor fix to consider `kCompactionStyleiLevel` as a compaction policy

+++ db/compaction/compaction_picker_ilevel.cc
@@ -267,7 +267,7 @@ bool ILevelCompactionBuilder::PickFileToCompact() {
 
   if (start_level_ < ilevel_) {
     // if `start_level_` < ilevel, pick all the files.
-    for (int index = 0; index <= level_files.size(); index++) {
+    for (int index = 0; index <= static_cast<int>(level_files.size()); index++) {
       auto* f = level_files[index];
       start_level_inputs_.files.push_back(f);
     }
